Switched to decimals Added basic operations tests

+++ Mathema.Interfaces/IExpression.cs
@@ -4,6 +4,6 @@ namespace Mathema.Interfaces
 {
     public interface IExpression
     {
-        double Value();
+        decimal Value();
     }
 }
