+++ Mathema.Models/Expressions/BinaryExpression.cs
@@ -19,7 +19,7 @@ namespace Mathema.Models.Expressions
             this.op = op;
         }
 
-        public double Value()
+        public decimal Value()
         {
             return Operations.BinaryOperations[op](lhe, rhe);
         }
